Extract search panel from search view
