+++ src/router/index.ts
@@ -13,7 +13,7 @@ const router = createRouter({
     {
       path: "/search",
       name: "search",
-      component: () => import("../views/SearchView.vue"),
+      component: () => import("../views/search/SearchView.vue"),
     },
     {
       path: "/browse",
